Move fixtures out of tests directory
