+++ packages/destination-actions/src/destinations/blackbaud-raisers-edge-nxt/createOrUpdateIndividualConstituent/__tests__/index.test.ts
@@ -18,7 +18,7 @@ import {
   identifyEventDataUpdated,
   onlinePresencePayloadUpdated,
   phonePayloadUpdated
-} from './fixtures'
+} from '../fixtures'
 
 const testDestination = createTestIntegration(Destination)
